fix sse connection
